backend/requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.1.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
numpy==2.1.1
pandas==2.2.3
pillow==10.4.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
